getting noise on the display

+++ src/lcd.h
@@ -7,7 +7,7 @@
 void LCD_IO_Init();
 void LCD_IO_WriteMultipleData(u8 *pData, u32 Size);
 void LCD_IO_WriteReg(u8 Reg);
-void LCD_IO_WriteData(u8 RegValue);
+void LCD_IO_WriteData(u16 RegValue);
 u16 LCD_IO_ReadData();
 void LCD_IO_Delay(u32 delay);
